fix: 게시글 도메인의 기존 DTO 클래스 수정
